+++ lib/src/main/java/org/jevon/gdx/logging/slf4j/package-info.java
@@ -1,0 +1,4 @@
+@NonNullByDefault
+package org.jevon.gdx.logging.slf4j;
+
+import org.eclipse.jdt.annotation.NonNullByDefault;